requirements.txt
discord.py
numpy
pillow
webp
git+https://github.com/martmists/mart_music#egg-info="async"
git+https://github.com/izunadevs/izunadsp
